Separated node dependencies publishing into its own installation command
